Mansi_and_Pragathi/requirements.txt
streamlit>=1.25
pandas>=1.5
google-generativeai
openpyxl
xlrd>=2.0.1
plotly>=5.0
scikit-learn
seaborn
numpy
anthropic
statsmodels